+++ src/buildWorkshop/config/BoostMultipliers.ts
@@ -1,3 +1,15 @@
 export const CLICK_BOOST_MULTIPLIER = 5;
 export const RESEARCH_BOOST_MULTIPLIER = 10;
 export const MERCHANT_BOOST_MULTIPLIER = 5;
+
+// TODO: GH-8: handle dynasty bonuses better
+export const DAILY_DYNASTY_FRIEND_BONUS_INCOME = 30;
+export const DAILY_DYNASTY_FRIEND_BONUS_MERCHANT = 30;
+export const DAILY_DYNASTY_FRIEND_BONUS_ORE = 10;
+
+export const PROMOTION_BONUS_INCOME = 1;
+export const PROMOTION_BONUS_MERCHANT = 1;
+export const PROMOTION_BONUS_RESEARCH = 2;
+export const PROMOTION_BONUS_CLICK_OUTPUT = 1;
+export const PROMOTION_BONUS_SPEED = 1;
+export const PROMOTION_BONUS_LPP = 0;
