+++ src/buildWorkshop/constants/Achievements.ts
@@ -8,15 +8,3 @@ export const MWS_EVENT_ACHIEVE_INCOME_MULTIPLIER = 6;
 
 export const RESEARCH_ACHIEVEMENT_MULTIPLIER = 10;
 export const TOTAL_BLUEPRINT_SCORE_MULTIPLIER = 2.8;
-
-// TODO: GH-8: handle dynasty bonuses better
-export const DAILY_DYNASTY_FRIEND_BONUS_INCOME = 30;
-export const DAILY_DYNASTY_FRIEND_BONUS_MERCHANT = 30;
-export const DAILY_DYNASTY_FRIEND_BONUS_ORE = 10;
-
-export const PROMOTION_BONUS_INCOME = 1;
-export const PROMOTION_BONUS_MERCHANT = 1;
-export const PROMOTION_BONUS_RESEARCH = 2;
-export const PROMOTION_BONUS_CLICK_OUTPUT = 1;
-export const PROMOTION_BONUS_SPEED = 1;
-export const PROMOTION_BONUS_LPP = 0;
