+++ src/buildWorkshop/helpers/ResearchHelpers.ts
@@ -1,12 +1,8 @@
 import memoize from 'fast-memoize';
 import { SetMultiplierType } from '../../upgradeBlueprints/constants/BlueprintSets';
 import { getSpecifiedMultiplierFromLibrary } from '../../upgradeBlueprints/helpers/blueprintScoreHelpers';
-import { RESEARCH_BOOST_MULTIPLIER } from '../config/BoostMultipliers';
-import {
-  PROMOTION_BONUS_RESEARCH,
-  RESEARCH_ACHIEVEMENT_MULTIPLIER,
-  TOTAL_BLUEPRINT_SCORE_MULTIPLIER,
-} from '../constants/Achievements';
+import { PROMOTION_BONUS_RESEARCH, RESEARCH_BOOST_MULTIPLIER } from '../config/BoostMultipliers';
+import { RESEARCH_ACHIEVEMENT_MULTIPLIER, TOTAL_BLUEPRINT_SCORE_MULTIPLIER } from '../constants/Achievements';
 import { Workshop, WorkshopStatus } from '../types/Workshop';
 import { isEvent } from './WorkshopHelpers';
